move theme folder into Service
