+++ app-test/basic/breadcrumb.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Fohn\Ui\AppTest;
 
+use Fohn\Ui\Service\Theme\Fohn;
 use Fohn\Ui\Service\Ui;
-use Fohn\Ui\Tailwind\Theme\Fohn;
 use Fohn\Ui\View\Breadcrumb;
 
 require_once __DIR__ . '/../init-ui.php';
