+++ app-test/javascript/js.php
@@ -9,9 +9,9 @@ use Fohn\Ui\Js\Js;
 use Fohn\Ui\Js\JsFunction;
 use Fohn\Ui\Js\JsReload;
 use Fohn\Ui\Js\Type\Variable;
+use Fohn\Ui\Service\Theme\Base;
+use Fohn\Ui\Service\Theme\Fohn;
 use Fohn\Ui\Service\Ui;
-use Fohn\Ui\Tailwind\Theme\Base;
-use Fohn\Ui\Tailwind\Theme\Fohn;
 use Fohn\Ui\View;
 use Fohn\Ui\View\Button;
 use Fohn\Ui\View\Heading\Header;
