+++ app-test/tw-default.php
@@ -7,7 +7,7 @@ declare(strict_types=1);
  */
 require_once __DIR__ . '/../vendor/autoload.php';
 
-use Fohn\Ui\Tailwind\Theme\Fohn;
+use Fohn\Ui\Service\Theme\Fohn;
 
 file_put_contents('fohn-theme-tw.txt', Fohn::getThemeCss());
 echo 'theme css done!' . \PHP_EOL;
