+++ src/Service/Theme/Base.php
@@ -6,7 +6,7 @@ declare(strict_types=1);
  * Base theme configuration.
  */
 
-namespace Fohn\Ui\Tailwind\Theme;
+namespace Fohn\Ui\Service\Theme;
 
 use Fohn\Ui\Tailwind\Tw;
 use Fohn\Ui\Tailwind\Utilities\TwDefaults;
