+++ src/Service/Ui.php
@@ -21,15 +21,14 @@ use Fohn\Ui\Core\Exception;
 use Fohn\Ui\Core\ExceptionRenderer\Html;
 use Fohn\Ui\Core\Utils;
 use Fohn\Ui\HtmlTemplate;
-use Fohn\Ui\Js\Js;
 use Fohn\Ui\Js\JsChain;
 use Fohn\Ui\Js\JsRenderInterface;
 use Fohn\Ui\Js\Type\Type;
 use Fohn\Ui\Page;
 use Fohn\Ui\PageLayout\Layout;
-use Fohn\Ui\Tailwind\Theme\Base;
-use Fohn\Ui\Tailwind\Theme\Fohn;
-use Fohn\Ui\Tailwind\Theme\ThemeInterface;
+use Fohn\Ui\Service\Theme\Base;
+use Fohn\Ui\Service\Theme\Fohn;
+use Fohn\Ui\Service\Theme\ThemeInterface;
 use Fohn\Ui\View;
 use Fohn\Ui\ViewRenderer;
 use GuzzleHttp\Psr7\Query;
