+++ src/View/Button.php
@@ -9,8 +9,8 @@ namespace Fohn\Ui\View;
 use Fohn\Ui\Js\Jquery;
 use Fohn\Ui\Js\Js;
 use Fohn\Ui\Js\JsRenderInterface;
+use Fohn\Ui\Service\Theme\Base;
 use Fohn\Ui\Service\Ui;
-use Fohn\Ui\Tailwind\Theme\Base;
 use Fohn\Ui\Tailwind\Tw;
 use Fohn\Ui\View;
