+++ src/View/Chip.php
@@ -8,8 +8,8 @@ declare(strict_types=1);
 
 namespace Fohn\Ui\View;
 
+use Fohn\Ui\Service\Theme\Base;
 use Fohn\Ui\Service\Ui;
-use Fohn\Ui\Tailwind\Theme\Base;
 use Fohn\Ui\Tailwind\Tw;
 use Fohn\Ui\View;
